backend/requirements.txt
flask
fastapi 
uvicorn
scikit-learn
pandas 
numpy 
tensorflow 
xgboost 
lightgbm 
kafka-python 
pymongo 
psycopg2-binary
joblib